# Conflicts:
#	README.md
#	WowPacketParser/Enums/ClientVersionBuild.cs
#	WowPacketParser/Enums/Version/Opcodes.cs
#	WowPacketParser/Enums/Version/UpdateFields.cs
#	WowPacketParser/Misc/ClientVersion.cs
#	WowPacketParser/Store/Objects/Creature.cs
#	WowPacketParser/Store/Objects/GossipMenu.cs
#	WowPacketParser/WowPacketParser.csproj
#	WowPacketParserModule.V6_0_2_19033/Parsers/AchievementHandler.cs
#	WowPacketParserModule.V7_0_3_22248/Parsers/InstanceHandler.cs
#	WowPacketParserModule.V7_0_3_22248/Parsers/QuestHandler.cs
#	WowPacketParserModule.V7_0_3_22248/Parsers/SpellHandler.cs

+++ WowPacketParserModule.V7_2_0_23758/Parsers/QuestHandler.cs
@@ -277,7 +277,7 @@ namespace WowPacketParserModule.V7_2_0_23758.Parsers
             quest.AreaGroupID = packet.ReadUInt32("AreaGroupID");
             quest.TimeAllowed = packet.ReadUInt32("TimeAllowed");
             uint int2946 = packet.ReadUInt32("CliQuestInfoObjective");
-            quest.AllowableRacesWod = packet.ReadInt32("AllowableRaces");
+            quest.AllowableRacesWod = (uint)packet.ReadInt32("AllowableRaces");
             quest.QuestRewardID = packet.ReadInt32("QuestRewardID");
             packet.ReadInt32("unk");
